[REFACTOR] rename utils.h moved QueryScore to different file
