+++ src/functions.cpp
@@ -1,5 +1,5 @@
 
-#include "functions.h"
+#include "utils.h"
 
 
 std::string str_join( const std::vector<std::string>& elements, const char* const separator)
